Add Simple Render System
