+++ src/first_app.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "lve_window.hpp"
-#include "lve_pipeline.hpp"
 #include "lve_device.hpp"
 #include "lve_game_object.hpp"
 #include "lve_renderer.hpp"
@@ -25,9 +24,6 @@ namespace lve {
             void run();
         private:
             void loadGameObjects();
-            void createPipelineLayout();
-            void createPipeline();
-            void renderGameObjects(VkCommandBuffer commandBuffer);
 
             // optional fun
             void seirpinskiSieve(float x, float y, float length, uint32_t iter, std::vector<LveModel::Vertex> &vertices);
@@ -35,8 +31,6 @@ namespace lve {
             LveWindow lveWindow{WIDTH, HEIGHT, "Hello, Vulkan!"};
             LveDevice lveDevice{lveWindow};
             LveRenderer lveRenderer{lveWindow, lveDevice};
-            std::unique_ptr<LvePipeline> lvePipeline;
-            VkPipelineLayout pipelineLayout;
             std::vector<LveGameObject> gameObjects;
     };
 }
