+++ src/lve_renderer.hpp
@@ -43,6 +43,6 @@ namespace lve {
             std::vector<VkCommandBuffer> commandBuffers;
 
             uint32_t currentImageIndex;
-            bool isFrameStarted;
+            bool isFrameStarted{false};
     };
 }
