+++ src/simple_render_system.cpp
@@ -1,0 +1,95 @@
+#include "simple_render_system.hpp"
+
+// Libs
+#define GLM_FORCE_RADIANS
+#define GLM_FORCE_DEPTH_ZERO_TO_ONE
+#include <glm/glm.hpp>
+#include <glm/gtc/constants.hpp>
+
+// Std
+#include <stdexcept>
+#include <memory>
+#include <array>
+#include <iostream>
+
+namespace lve {
+
+    struct SimplePushConstantData {
+        glm::mat2 transform{1.0f};
+        glm::vec2 offset;
+        alignas(16) glm::vec3 color;
+    };
+
+    SimpleRenderSystem::SimpleRenderSystem(LveDevice &device, VkRenderPass renderPass) 
+        : lveDevice{device} {
+        createPipelineLayout();
+        createPipeline(renderPass);
+    }
+
+    SimpleRenderSystem::~SimpleRenderSystem() {
+        vkDestroyPipelineLayout(lveDevice.device(), pipelineLayout, nullptr);
+    }
+
+    void SimpleRenderSystem::createPipelineLayout() {
+        std::cout << "Creating Pipeline Layout...\n";
+
+        VkPushConstantRange pushConstantRange{};
+        pushConstantRange.stageFlags = VK_SHADER_STAGE_VERTEX_BIT | VK_SHADER_STAGE_FRAGMENT_BIT;
+        pushConstantRange.offset = 0;
+        pushConstantRange.size = sizeof(SimplePushConstantData);
+
+        VkPipelineLayoutCreateInfo pipelineLayoutInfo{};
+        pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
+        pipelineLayoutInfo.setLayoutCount = 0;
+        pipelineLayoutInfo.pSetLayouts = nullptr;
+        pipelineLayoutInfo.pushConstantRangeCount = 1;
+        pipelineLayoutInfo.pPushConstantRanges = &pushConstantRange;
+
+        if (vkCreatePipelineLayout(lveDevice.device(), &pipelineLayoutInfo, nullptr, &pipelineLayout) != VK_SUCCESS) {
+            throw std::runtime_error("Failed to create pipeline layout");
+        } 
+    }
+
+    void SimpleRenderSystem::createPipeline(VkRenderPass renderPass) {
+        std::cout << "Creating Pipeline...\n";
+        assert(pipelineLayout != nullptr && "Cannot crate pipeline before pipeline layout!");
+
+        PipelineConfigInfo pipelineConfig{};
+        LvePipeline::defaultPipelineConfigInfo(pipelineConfig);
+        pipelineConfig.renderPass = renderPass;
+        pipelineConfig.pipelineLayout = pipelineLayout;
+        lvePipeline = std::make_unique<LvePipeline>(
+            lveDevice,
+            "shaders/simple_shader.vert.spv",
+            "shaders/simple_shader.frag.spv",
+            pipelineConfig
+        );
+
+        std::cout << "End Create Pipeline...\n";
+    }
+
+    void SimpleRenderSystem::renderGameObjects(VkCommandBuffer commandBuffer, std::vector<LveGameObject> &gameObjects) {
+        lvePipeline->bind(commandBuffer);
+
+        for (auto& object: gameObjects) {
+
+            object.transform2d.rotation = glm::mod(object.transform2d.rotation + 0.001f, glm::two_pi<float>());
+
+            SimplePushConstantData push{};
+            push.offset = object.transform2d.translation;
+            push.color = object.color;
+            push.transform = object.transform2d.mat2();
+
+            vkCmdPushConstants(
+                commandBuffer,
+                pipelineLayout, 
+                VK_SHADER_STAGE_VERTEX_BIT | VK_SHADER_STAGE_FRAGMENT_BIT,
+                0,
+                sizeof(SimplePushConstantData),
+                &push);
+
+            object.model->bind(commandBuffer);
+            object.model->draw(commandBuffer);
+        }
+    }
+}
